refactored `rendering.py` to enable switching between osmesa and egl during runtime, added equivalence test for the two rendering platforms, added new syntype model
