+++ syconn/cnn/cnn_synapse_type_enhanced.py
@@ -14,12 +14,54 @@ import torch
 from torch import nn
 from torch import optim
 
-# TODO: Make torch and numpy RNG seed configurable
 
+class HybridDiceLoss(torch.nn.Module):
+    def __init__(self):
+        super().__init__()
+        self.dice_syntype = DiceLossFancy(
+            apply_softmax=True, weights=torch.tensor([0.33, 0.33, 0.33, 0]).to(device),
+            ignore_index=3)
+        weights = torch.tensor([1.] * 10 + [0., ]).to(device)
+        self.dice_celltype = DiceLossFancy(
+            apply_softmax=True, weights=weights, ignore_index=10)
+        self.dice_celltype2 = DiceLossFancy(
+            apply_softmax=True, weights=weights, ignore_index=10)
+        self.mse = torch.nn.MSELoss()  # use `reduce` kwarg?
+
+    def forward(self, output, target, downscale_fact=0.1):
+        # target shape: B, C, Z, Y, X
+        # output shape: B, OUT_C, Z, Y, X
+        # C: vector field 0-2, syntype label 3, celltype labels 4-5
+        # OUT_C: vec. field 0-2, syntype 3-6, celltype 7-17 and 18-28
+        vec_field_d = output[:, :3]
+        vec_field_t = target[:, :3]
+        # scale target vectors
+        loss_vec = self.mse(vec_field_d, vec_field_t.to(torch.float32) * downscale_fact)
+        syntype_d = output[:, 3:7]
+        syntype_l = target[:, 3]
+        loss_syntype = self.dice_syntype(syntype_d, syntype_l)
+        celltype_d = output[:, 7:18]
+        celltype_l = target[:, 4]
+        loss_celltype = self.dice_celltype(celltype_d, celltype_l)
+        celltype_d2 = output[:, 18:]
+        celltype_l2 = target[:, 5]
+        loss_celltype2 = self.dice_celltype2(celltype_d2, celltype_l2)
+        if torch.isnan(loss_vec):
+            raise ValueError('Vectorial loss is NaN.')
+        if torch.isnan(loss_syntype):
+            raise ValueError('Synapsetype loss is NaN.')
+        if torch.isnan(loss_celltype):
+            raise ValueError('Celltype loss is NaN.')
+        if torch.isnan(loss_celltype2):
+            raise ValueError('Celltype 2 loss is NaN.')
+        return loss_vec + loss_syntype + loss_celltype + loss_celltype2
+
+
+# TODO: Make torch and numpy RNG seed configurable
 parser = argparse.ArgumentParser(description='Train a network.')
 parser.add_argument('--disable-cuda', action='store_true', help='Disable CUDA')
 parser.add_argument('-n', '--exp-name',
-                    default='syntype_unet_sameConv_noBN_fancydice_gt3_bs4_inmem_run2',
+                    default='syntype_unet_sameConv_BN_fancydice_gt3_enhanced_bs4_inmem_10xmsefact',
                     help='Manually set experiment name')
 parser.add_argument(
     '-s', '--epoch-size', type=int, default=500,
@@ -64,32 +106,17 @@ from elektronn3.data import PatchCreator, transforms, utils, get_preview_batch
 from elektronn3.training import Trainer, Backup, metrics, Padam, handlers
 from elektronn3.models.unet import UNet
 from elektronn3.modules.loss import DiceLoss, DiceLossFancy
-from elektronn3.training.metrics import channel_metric
 
 torch.backends.cudnn.benchmark = True  # Improves overall performance in *most* cases
 
-# model = UNet_valid_blocks(
-#     out_channels=5,
-#     n_blocks=5,
-#     start_filts=64,
-#     planar_blocks=(1,2),
-#     activation='relu',
-#     batch_norm=True,
-#     valid_blocks = (0,1,2),
-#     adaptive=False
-# ).to(device)
-# offset = model.offset
-
 model = UNet(
     in_channels=1,
-    out_channels=4,
+    out_channels=29,  # vec. field 0-2, syntype 3-6, celltype 7-17 and 18-28
     n_blocks=4,
     start_filts=28,
     planar_blocks=(0,),
     activation='relu',
-    batch_norm=False,
-    # conv_mode='valid',
-    #up_mode='resizeconv_nearest',  # Enable to avoid checkerboard artifacts
+    batch_norm=True,
     adaptive=False  # Experimental. Disable if results look weird.
 ).to(device)
 
@@ -109,28 +136,22 @@ elif args.jit == 'train':
     tracedmodel = torch.jit.trace(model, example_input.to(device))
     model = tracedmodel
 
-
 # USER PATHS
 save_root = os.path.expanduser('~/e3_training/')
 os.makedirs(save_root, exist_ok=True)
 data_root = os.path.expanduser('/ssdscratch/pschuber/songbird/j0126/GT/synapsetype_gt/')
 
-gt_dir = data_root + '/Segmentierung_von_Synapsentypen_v3/'
-fnames = sorted([gt_dir + f for f in os.listdir(gt_dir) if f.endswith('.h5')])
-gt_dir = data_root + '/synssv_reconnects_nosomamerger/'
+gt_dir = data_root + '/synssv_reconnects_nosomamerger_enhanced/'
 fnames_files = sorted([gt_dir + f for f in os.listdir(gt_dir) if f.endswith('.h5')])
 random_ixs = np.arange(len(fnames_files))
 np.random.seed(0)
 np.random.shuffle(fnames_files)
 fnames_files = np.array(fnames_files)[random_ixs].tolist()
-fnames += fnames_files[:900]
+fnames = fnames_files[:950]
 
-input_h5data = [(f, 'raw') for f in fnames + fnames[-1:]]
-target_h5data = [(f, 'label') for f in fnames + fnames[-1:]]
-valid_indices = [len(target_h5data) - 1]
-
-# Class weights for imbalanced dataset, the last one is used as ignore label
-class_weights = torch.tensor([0.33, 0.33, 0.33, 0]).to(device)
+input_h5data = [(f, 'raw') for f in fnames + fnames[-5:]]
+target_h5data = [(f, 'label') for f in fnames + fnames[-5:]]
+valid_indices = np.arange(len(target_h5data) - 5, len(target_h5data))
 
 max_steps = args.max_steps
 max_runtime = args.max_runtime
@@ -145,7 +166,6 @@ if args.resume is not None:  # Load pretrained network
 drop_func = transforms.DropIfTooMuchBG(bg_id=3, threshold=0.9)
 # Transformations to be applied to samples before feeding them to the network
 common_transforms = [
-    transforms.SqueezeTarget(dim=0),  # Workaround for neuro_data_cdhw
     #transforms.Normalize(mean=dataset_mean, std=dataset_std),
 ]
 train_transform = transforms.Compose(common_transforms + [
@@ -162,14 +182,10 @@ aniso_factor = 2  # Anisotropy in z dimension. E.g. 2 means half resolution in z
 common_data_kwargs = {  # Common options for training and valid sets.
     'aniso_factor': aniso_factor,
     'patch_shape': (48, 144, 144),
-    'num_classes': 4,
+    'num_classes': 6,
     # 'offset': (20, 46, 46),
+    'target_discrete_ix': [3, 4, 5]
 }
-
-if len(fnames) > 100:  # artificial GT based on cell types is in use
-    cube_prios = None
-else:
-    cube_prios = [1] * (len(input_h5data) - len(valid_indices))
 
 type_args = list(range(len(input_h5data)))
 train_dataset = PatchCreator(
@@ -177,7 +193,7 @@ train_dataset = PatchCreator(
     target_h5data=[target_h5data[i] for i in type_args if i not in valid_indices],
     train=True,
     epoch_size=args.epoch_size,
-    cube_prios=cube_prios,
+    cube_prios=None,
     warp_prob=0.2,
     in_memory=True,
     warp_kwargs={
@@ -188,7 +204,7 @@ train_dataset = PatchCreator(
     transform=train_transform,
     **common_data_kwargs
 )
-valid_dataset = None if not valid_indices else PatchCreator(
+valid_dataset = None if (valid_indices is None) or (len(valid_indices) == 0) else PatchCreator(
     input_h5data=[input_h5data[i] for i in range(len(input_h5data)) if i in valid_indices],
     target_h5data=[target_h5data[i] for i in range(len(input_h5data)) if i in valid_indices],
     train=False,
@@ -222,8 +238,7 @@ schedulers = {'lr': lr_sched}
 valid_metrics = {
 }
 
-criterion = DiceLossFancy(apply_softmax=True, weights=class_weights,
-                          ignore_index=3)
+criterion = HybridDiceLoss()
 
 # Create trainer
 trainer = Trainer(
@@ -239,30 +254,18 @@ trainer = Trainer(
     exp_name=args.exp_name,
     example_input=example_input,
     enable_save_trace=enable_save_trace,
-    schedulers=schedulers,#{"lr": optim.lr_scheduler.StepLR(optimizer, 1000, 0.995)},
+    schedulers=schedulers,
     valid_metrics=valid_metrics,
-    #preview_batch=preview_batch,
-    #preview_interval=5,
     enable_videos=False,  # Uncomment to get rid of videos in tensorboard
     offset=train_dataset.offset,
     apply_softmax_for_prediction=True,
     num_classes=train_dataset.num_classes,
     ipython_shell=False,
-    # TODO: Tune these:
-    #preview_tile_shape=(48, 96, 96),
-    #preview_overlap_shape=(48, 96, 96),
-    #sample_plotting_handler = handlers._tb_log_sample_images_Synapse,
-    #mixed_precision=True,  # Enable to use Apex for mixed precision training
+
 )
 
 # Archiving training script, src folder, env info
-Backup(script_path=__file__,save_path=trainer.save_path).archive_backup()
+Backup(script_path=__file__, save_path=trainer.save_path).archive_backup()
 
 # Start training
 trainer.run(max_steps=max_steps, max_runtime=max_runtime)
-
-
-# How to re-calculate mean, std and class_weights for other datasets:
-#  dataset_mean = utils.calculate_means(train_dataset.inputs)
-#  dataset_std = utils.calculate_stds(train_dataset.inputs)
-#  class_weights = torch.tensor(utils.calculate_class_weights(train_dataset.targets))
