+++ syconn/proc/meshes.py
@@ -57,7 +57,7 @@ except ImportError:
 __all__ = ['MeshObject', 'get_object_mesh', 'merge_meshes', 'triangulation',
            'get_random_centered_coords', 'write_mesh2kzip', 'write_meshes2kzip',
            'compartmentalize_mesh', 'mesh_chunk', 'mesh_creator_sso', 'merge_meshes_incl_norm',
-           'mesh_area_calc', 'mesh2obj_file']
+           'mesh_area_calc', 'mesh2obj_file', 'calc_rot_matrices']
 
 
 class MeshObject(object):
